requirements.txt
llvmlite==0.31.0
numba==0.48.0
numpy==1.18.2
pygame==1.9.6
scipy==1.4.1
pyglet==1.5.5
pillow==6.2.1